Fix for focus issue
When clicked on, was focusing the canvas element instead of the window. This was causing the keyboard events to be swallowed by the canvas instead of the window. Fix was to add the keyboard listener to both elements

Also right now keyboard event logging is in verbose mode so its spamming the sys out...